Created methods to generate the XML for a content and script editor webpart.
