+++ src/helper/methods/addScriptEditorWebPart.ts
@@ -1,4 +1,5 @@
 import { IScriptEditorWebPart } from "../../../@types/helper";
+import { WebPart } from "../webpart";
 declare var SP;
 
 // Method to add a script editor webpart to the page
@@ -13,28 +14,7 @@ export const addScriptEditorWebPart = (url: string, wpProps: IScriptEditorWebPar
         let wpMgr = page.getLimitedWebPartManager(SP.WebParts.PersonalizationScope.shared);
 
         // Create the webpart
-        let wp = wpMgr.importWebPart(`<?xml version="1.0" encoding="utf-8"?>
-<webParts>
-    <webPart xmlns="http://schemas.microsoft.com/WebPart/v3">
-        <metaData>
-            <type name="Microsoft.SharePoint.WebPartPages.ScriptEditorWebPart, Microsoft.SharePoint, Version=15.0.0.0, Culture=neutral, PublicKeyToken=71e9bce111e9429c" />
-            <importErrorMessage>$Resources:core,ImportantErrorMessage;</importErrorMessage>
-        </metaData>
-        <data>
-            <properties>
-            <property name="Title" type="string">[[Title]]</property>
-            <property name="Description" type="string">[[Description]]</property>
-            <property name="ChromeType" type="chrometype">[[ChromeType]]</property>
-            <property name="Content" type="string">[[Content]]</property>
-        </properties>
-        </data>
-    </webPart>
-</webParts>`.replace(/\r?\n/g, '')
-            .replace(/\[\[ChromeType\]\]/g, wpProps.chromeType || "TitleOnly")
-            .replace(/\[\[Content\]\]/g, wpProps.content.replace(/\</g, '&lt;').replace(/\>/g, '&gt;'))
-            .replace(/\[\[Description\]\]/g, wpProps.description || "")
-            .replace(/\[\[Title\]\]/g, wpProps.title || "")
-        ).get_webPart();
+        let wp = wpMgr.importWebPart(WebPart.generateScriptEditorXML(wpProps)).get_webPart();
 
         // Add the webpart to the page
         wpMgr.addWebPart(wp, wpProps.zone || "", wpProps.index || 0);
